update jdbc connector and update readme
