--- docker/wildfly/jdbc/mysql-connector-java-5.1.37-bin.jar.index
@@ -1,32 +1,0 @@
-
-com/mysql/jdbc/authentication
-com
-com/mysql/jdbc
-com/mysql/fabric/jdbc
-com/mysql/fabric/hibernate
-com/mysql/jdbc/jdbc2
-com/mysql/fabric/xmlrpc
-org/gjt/mm
-com/mysql/jdbc/profiler
-com/mysql/fabric/proto/xmlrpc
-org
-com/mysql/jdbc/jmx
-com/mysql/jdbc/integration/c3p0
-com/mysql/jdbc/exceptions
-com/mysql/jdbc/interceptors
-com/mysql/jdbc/util
-com/mysql/jdbc/exceptions/jdbc4
-com/mysql/fabric/proto
-com/mysql/fabric/xmlrpc/exceptions
-com/mysql/fabric/xmlrpc/base
-org/gjt/mm/mysql
-com/mysql/jdbc/log
-com/mysql/fabric
-com/mysql/jdbc/integration/jboss
-org/gjt
-META-INF
-com/mysql/jdbc/configs
-com/mysql/jdbc/jdbc2/optional
-com/mysql
-com/mysql/jdbc/integration
-META-INF/services
