+++ src/main/java/dao/AbstractDao.java
@@ -2,13 +2,10 @@ package dao;
 
 import java.io.Serializable;
 import java.lang.reflect.ParameterizedType;
- 
+
 import javax.persistence.EntityManager;
 import javax.persistence.PersistenceContext;
 
-import org.springframework.stereotype.Component;
-
-@Component
 public abstract class AbstractDao<PK extends Serializable, T> {
      
     private final Class<T> persistentClass;
